audio: Move IcySrc out of playlist module
