+++ mopidy/audio/actor.py
@@ -13,7 +13,7 @@ import gst.pbutils  # noqa
 import pykka
 
 from mopidy import exceptions
-from mopidy.audio import playlists, utils
+from mopidy.audio import icy, playlists, utils
 from mopidy.audio.constants import PlaybackState
 from mopidy.audio.listener import AudioListener
 from mopidy.internal import deprecation, process
@@ -26,6 +26,7 @@ logger = logging.getLogger(__name__)
 # set_state on a pipeline.
 gst_logger = logging.getLogger('mopidy.audio.gst')
 
+icy.register()
 playlists.register_typefinders()
 playlists.register_elements()
